chore: moved test to ContextBroker
For #148
